Updates to radiation to include ionization of HeII, HeI and dissociation of H2 (part 5)

+++ src/stars/star_radiation.h
@@ -78,7 +78,7 @@ typedef struct
 typedef struct {
   RayPacket *rays;
   int n;
-  int capacity;
+  long long capacity;
 } RayWorkStack;
 
 typedef struct
